Rename sw.js
