+++ src/app.js
@@ -5,7 +5,7 @@ import * as sidebar from './sidebar.js';
 import * as notesList from './notesList.js';
 import * as scrollbar from './scrollbar.js';
 import * as theme from './theme.js';
-import * as serviceWorker from './sw.js';
+import * as serviceWorker from './serviceWorker.js';
 import { removeClass } from './helpers.js';
 
 const body = document.querySelector('body');
